fix: use indicatiff from crates.io
Indicatif MultiProgress doesn't work well in the static context
This is basically a workaround for that.
A regular thread is spawned to update the progress in the background.

Note that the code might be refactored to the way it was before.
Once https://github.com/mitsuhiko/indicatif/pull/132 is merged.